Connection classes: Show debug output as pretty JSON
When debug is on, the connection classes always dumped messages sent and
received as compact .raku output.  Unfortunately for large, complex messages
it was nigh unusable for finding subtle structural bugs.

Now with debug on, the connection classes show client ↔ server message passing
as pretty-printed JSON.  The in-process Supplier connection classes still show
in-client and in-server message passing as .raku output in order to catch e.g.
an Array replaced with a Seq.