Get rid of spurious WebSocket Continuation frames
Previously, every single WebSocket message was sent "fragmented", which means
that after each data-containing frame, a second empty "continuation" frame was
also sent, just to mark the end of the fragmented message.

It turns out this was the result of an unfortunate DWIM attempt in
Cro::WebSocket::Message, specifically with this constructor:

https://github.com/croservices/cro-websocket/blob/master/lib/Cro/WebSocket/Message.pm6#L28

If sending general data in the message body (to be encoded into binary with
Cro::CBOR), that constructor would guess that the message should be fragmented
because it was neither a Str nor a Blob *before encoding it*.

The workaround is to simply wrap the general data to be sent in a
Cro::WebSocket::Message manually, explicitly turning off the fragmented flag.

This results in about HALF the WebSocket frame rate in both directions, and
avoids a lot of wasted serialization and deserialization of empty continuation
frames in the Cro::WebSocket stack as well.

Longer term perhaps Cro::Websocket should be fixed to determine if the message
should be fragmented *after* serializing/encoding the body and thereby knowing
what type it will actually be, but there may be other existing Cro apps that
depend on the current behavior, so this is probably an upstream decision.